makes sample queries filterable
