+++ tslint.json
@@ -92,7 +92,7 @@
   "linterOptions": {
     "exclude": [
       "src/messages/*",
-      "src/app/views/sidebar/sample-queries.ts"
+      "src/app/views/sidebar/sample-queries/queries.ts"
     ]
   }
 }
